Master

+++ Back/Test.Domain/Entities/AsignaturaAlumno.cs
@@ -1,5 +1,7 @@
 ﻿ 
  
+ using System;
+
  namespace Test.Domain.Entities
 {
     public class AsignaturaAlumno : Entity
@@ -7,7 +9,7 @@
         public int IdAlumno { get; set; }
         public int IdAsignatura { get; set; }
         public short AnioLectivo { get; set; }
-        public byte Calificacion { get; set; }
+        public decimal Calificacion { get; set; }
 
         public  Alumno IdAlumnoNavigation { get; set; }
         public  Asignatura IdAsignaturaNavigation { get; set; }
